requirements.txt
aiofiles==23.2.1
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.3.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclass-factory==2.16
ffmpeg-python==0.2.0
frozenlist==1.5.0
future==1.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.2.0
numpy==2.2.2
propcache==0.3.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
python-dotenv==1.0.1
python-telegram-bot==22.0
requests==2.32.3
shazamio==0.8.0
shazamio_core==1.1.2
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
yt-dlp==2025.2.19